refactor scope change components
