+++ src/apps/ScopeChangeRequest/Components/Workflow/Criteria/Components/WorkflowCriteria/CriteriaDetail/CriteriaDetail.tsx
@@ -1,8 +1,12 @@
-import styled from 'styled-components';
-import { Criteria, CriteriaSignState, WorkflowStep } from '../../../../types/scopeChangeRequest';
-import { convertUtcToLocalDate, dateToDateTimeFormat } from '../../Utils/dateFormatting';
-import { WorkflowIcon } from '../../Components/WorkflowIcon';
-import { tokens } from '@equinor/eds-tokens';
+import {
+    Criteria,
+    CriteriaSignState,
+    WorkflowStep,
+} from '../../../../../../types/scopeChangeRequest';
+import { convertUtcToLocalDate, dateToDateTimeFormat } from '../../../../Utils/dateFormatting';
+import { WorkflowIcon } from '../../../../Components/WorkflowIcon';
+import { DetailText, SplitInline, WorkflowText } from './criteriaDetail.styles';
+import { getCriteriaStatus } from './Utils/getCriteriaStatus';
 
 interface CriteriaDetailProps {
     criteria: Criteria;
@@ -11,19 +15,11 @@ interface CriteriaDetailProps {
 
 export type CriteriaStatus = CriteriaSignState | 'Inactive' | 'Active';
 
-function statusFunc(criteria: Criteria, step: WorkflowStep): CriteriaStatus {
-    if (!criteria.signedState) {
-        return step.isCurrent ? 'Active' : 'Inactive';
-    } else {
-        return criteria.signedState;
-    }
-}
-
 export const CriteriaDetail = ({ criteria, step }: CriteriaDetailProps): JSX.Element => {
     const date = convertUtcToLocalDate(new Date(criteria.signedAtUtc || new Date()));
     const formattedDate = dateToDateTimeFormat(date);
 
-    const stepStatus = statusFunc(criteria, step);
+    const stepStatus = getCriteriaStatus(criteria, step);
 
     return (
         <SplitInline>
@@ -43,22 +39,3 @@ export const CriteriaDetail = ({ criteria, step }: CriteriaDetailProps): JSX.Ele
         </SplitInline>
     );
 };
-
-const DetailText = styled.div`
-    font-size: 14px;
-`;
-
-const SplitInline = styled.div`
-    display: flex;
-    align-items: center;
-    justify-content: space-between;
-    gap: 1em;
-`;
-
-const WorkflowText = styled.div`
-    display: flex;
-    flex-direction: column;
-    align-items: flex-start;
-    font-size: 16px;
-    color: ${tokens.colors.text.static_icons__default.hex};
-`;
