+++ src/apps/ScopeChangeRequest/Components/Workflow/Criteria/index.ts
@@ -1,1 +1,1 @@
-export * from './Components/WorkflowCriteria/WorkflowCriteria'
+export * from './Components/WorkflowCriteria/WorkflowCriteria/WorkflowCriteria';
